+++ telegram/inline_query_result_location.go
@@ -9,10 +9,10 @@ type InlineQueryResultLocation struct {
 	Latitude            float64               `json:"latitude"`                        // Location latitude in degrees
 	Longitude           float64               `json:"longitude"`                       // Location longitude in degrees
 	Title               string                `json:"title"`                           // Location title
-	LivePeriod          int64                 `json:"live_period,omitempty"`           // Optional. Period in seconds for which the location can be updated, should be between 60 and 86400.
+	LivePeriod          int                   `json:"live_period,omitempty"`           // Optional. Period in seconds for which the location can be updated, should be between 60 and 86400.
 	ReplyMarkup         *InlineKeyboardMarkup `json:"reply_markup,omitempty"`          // Optional. Inline keyboard attached to the message
 	InputMessageContent *InputMessageContent  `json:"input_message_content,omitempty"` // Optional. Content of the message to be sent instead of the location
 	ThumbURL            string                `json:"thumb_url,omitempty"`             // Optional. URL of the thumbnail for the result
-	ThumbWidth          int64                 `json:"thumb_width,omitempty"`           // Optional. Thumbnail width
-	ThumbHeight         int64                 `json:"thumb_height,omitempty"`          // Optional. Thumbnail height
+	ThumbWidth          int                   `json:"thumb_width,omitempty"`           // Optional. Thumbnail width
+	ThumbHeight         int                   `json:"thumb_height,omitempty"`          // Optional. Thumbnail height
 }
