+++ telegram/invoice.go
@@ -8,5 +8,5 @@ type Invoice struct {
 	Description    string `json:"description"`     // Product description
 	StartParameter string `json:"start_parameter"` // Unique bot deep-linking parameter that can be used to generate this invoice
 	Currency       string `json:"currency"`        // Three-letter ISO 4217 currency code
-	TotalAmount    int64  `json:"total_amount"`    // Total price in the smallest units of the currency (integer, not float/double). For example, for a price of US$ 1.45 pass amount = 145. See the exp parameter in currencies.json, it shows the number of digits past the decimal point for each currency (2 for the majority of currencies).
+	TotalAmount    int    `json:"total_amount"`    // Total price in the smallest units of the currency (integer, not float/double). For example, for a price of US$ 1.45 pass amount = 145. See the exp parameter in currencies.json, it shows the number of digits past the decimal point for each currency (2 for the majority of currencies).
 }
